requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bootstrap4==0.1.0
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.2.0
django-braces==1.14.0
django-extensions==3.0.5
django-filter==2.3.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-rating==0.1.5
django-registration==3.1
django-registration-redux==2.8
django-star-ratings==0.9.1
djangorestframework==3.11.1
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0

psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pyuploadcare==2.7.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
